refactor(data-recomposition): lift compose/recomposition payload schema to data-recomposition-core
Extract RecompositionPayload, RecompositionNode, KIND, SCHEMA_VERSION,
MODE_DELTA, and MODE_SNAPSHOT out of RecompositionDataProducer.kt into
a new data-recomposition-core module so MCP clients and other connectors
can depend on the schema without pulling in the daemon-side registry,
the experimental Compose runtime CompositionObserver API, or the
desktop ImageComposeScene reflection facade.

RecompositionDataProductRegistry.KIND / .SCHEMA_VERSION / .MODE_DELTA /
.MODE_SNAPSHOT stay as forwarding constants pointing at
RecompositionProduct so existing in-tree call sites keep working.

Mirrors the core/connector split that fonts, strings, layoutinspector,
resources, a11y, history, theme, and wallpaper already use. With this
landed, every data-product feature has the same shape: pure-data core
+ daemon-side connector + (optional) extension-shape modules. No
behavior change.
